bot/main.py
import os
import discord
from discord.ext import commands
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()
DISCORD_TOKEN = os.getenv("DISCORD_TOKEN")

# Initialize the bot
intents = discord.Intents.default()
intents.message_content = True
bot = commands.Bot(command_prefix="!", intents=intents)

@bot.event
async def on_ready():
    print(f'Logged in as {bot.user.name}')

@bot.command(name='crown')
async def crown(ctx):
    """(Placeholder) Mint a CrownKey."""
    await ctx.send("CrownKey minting is not yet implemented.")

@bot.command(name='witness')
async def witness(ctx):
    """(Placeholder) Query Supabase for witness status."""
    await ctx.send("Witness status query is not yet implemented.")

@bot.slash_command(name="unfold", description="(Placeholder) Unfold a new vision.")
async def unfold(ctx):
    """(Placeholder) Hook for sigil/vision generation."""
    await ctx.respond("Vision unfolding is not yet implemented.")

if __name__ == "__main__":
    if DISCORD_TOKEN is None:
        print("Error: DISCORD_TOKEN not found in .env file.")
    else:
        bot.run(DISCORD_TOKEN)
