bot/requirements.txt
discord.py
python-dotenv
supabase-py
apscheduler
